Update: Automated the scripting to create manim files.
Using subprocess, created a pipeline that takes a query,
and a filename (to store the manim script, using manim.py,
currently), then generates an output file, named animation-
<uuid4>.

Caveat: The media files are generated wherever the script is
called from. Have to fix it such that all media files are
stored in the manim_llm_core directory.